Move Instagram comments helper functions to apify directory
